requirements.txt
Django==4.2.7
djangorestframework==3.14.0
python-telegram-bot==20.7
python-decouple==3.8
django-unfold==0.19.0
drf-yasg==1.21.7


